gdk_pixbuf_get_from_surface: restore old behavior
A previous commit changed this function to return a pixbuf that is
larger than the passed (width, height) when the surface has a device
scale > 1. It's easy enough to take care of a scaled surface before
calling this function.

Revert to the old behaviour, but change gdk_pixbuf_get_from_window() to
return a scaled version when the underlying surface is scaled. We need
this because there's no other way to get to the unscaled window
contents.

https://bugzilla.gnome.org/show_bug.cgi?id=757147