config.h.win32.in: Update for _lock_file()
The Visual Studio versions that we support supports locking functions in
their CRT, so support that to optimize things a bit.  Also update the
config.h.win32.in so that its entries are more in line with the ones in
the autootols builds, and make sure that we use UNIX line endings.