text view: Resepct caret-shape CSS property
When drawing the text cursor, consult the caret-shape property
to determine what kind of cursor to draw. To implement this, we
had to slightly change the way GtkTextLineDisplay is used.
Previously, GtkTextLayout would either fill in the block_cursor
rectangle or the cursors array, depending on what cursor shape
is needed. Now, we fill them both in and decide at render time
which one to use.